refactor RunningJobs out of GerritTrigger; add null check on trigger policy
