+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/hudsontrigger/GerritTrigger.java
@@ -146,7 +146,7 @@ public class GerritTrigger extends Trigger<Job> {
     public static final String JOB_ABORT = GerritTrigger.class.getName() + "_job_abort";
 
     //! Association between patches and the jobs that we're running for them
-    private transient RunningJobs runningJobs = new RunningJobs();
+    private transient RunningJobs runningJobs = new RunningJobs(this, this.job);
     //! This latch will be used to signal that the project list is ready for use.
     //! For static configuration, this will be ready immediately.
     //! For dynamic configuration, this will be ready after the first time that
@@ -784,7 +784,7 @@ public class GerritTrigger extends Trigger<Job> {
      */
     /*package*/ synchronized RunningJobs getRunningJobs() {
         if (runningJobs == null) {
-            runningJobs = new RunningJobs();
+            runningJobs = new RunningJobs(this, this.job);
         }
         return runningJobs;
     }
@@ -799,7 +799,7 @@ public class GerritTrigger extends Trigger<Job> {
         if (event instanceof ChangeBasedEvent) {
             IGerritHudsonTriggerConfig serverConfig = getServerConfig(event);
             if ((serverConfig != null && serverConfig.isGerritBuildCurrentPatchesOnly())
-                    || this.getBuildCancellationPolicy().isEnabled()) {
+                    || (this.getBuildCancellationPolicy() != null && this.getBuildCancellationPolicy().isEnabled())) {
                 getRunningJobs().remove((ChangeBasedEvent)event);
             }
         }
@@ -2331,236 +2331,7 @@ public class GerritTrigger extends Trigger<Job> {
         }
     }
 
-    /**
-     * Class for maintaining and synchronizing the runningJobs info.
-     * Association between patches and the jobs that we're running for them.
-     */
-    public class RunningJobs {
-        private final Set<GerritTriggeredEvent> runningJobs =
-                Collections.synchronizedSet(new HashSet<GerritTriggeredEvent>());
-
-        /**
-         * Called when trigger has cancellation policy associated with it.
-         *
-         *
-         * @param event event that is trigger builds
-         * @param jobName job name to match for specific cancellation
-         * @param policy policy to decide cancelling build or not
-         */
-        public void cancelTriggeredJob(ChangeBasedEvent event, String jobName, BuildCancellationPolicy policy)
-        {
-            if (policy == null || !policy.isEnabled() && (event instanceof ManualPatchsetCreated
-                    && !policy.isAbortManualPatchsets())) {
-               return;
-            }
-            this.cancelOutDatedEvents(event, policy, jobName);
-        }
-
-        /**
-         * Checks scheduled job and cancels current jobs if needed.
-         * I.e. cancelling the old build if configured to do so and removing and storing any references.
-         *
-         * @param event the event triggering a new build.
-         */
-        public void scheduled(ChangeBasedEvent event) {
-            IGerritHudsonTriggerConfig serverConfig = getServerConfig(event);
-            if (serverConfig == null) {
-                runningJobs.add(event);
-                return;
-            }
-
-            BuildCancellationPolicy serverBuildCurrentPatchesOnly = serverConfig.getBuildCurrentPatchesOnly();
-
-            if (!serverBuildCurrentPatchesOnly.isEnabled()
-                    || (event instanceof ManualPatchsetCreated
-                    && !serverBuildCurrentPatchesOnly.isAbortManualPatchsets())) {
-                runningJobs.add(event);
-                return;
-            }
-
-            this.cancelOutDatedEvents(event, serverBuildCurrentPatchesOnly, null);
-        }
-
-        /**
-         *
-         * @param event event to check for
-         * @param policy policy to determine cancellation of build for
-         * @param jobName job name parameter to consider; if null, assumes all builds
-         */
-        private void cancelOutDatedEvents(ChangeBasedEvent event, BuildCancellationPolicy policy, String jobName)
-        {
-            List<ChangeBasedEvent> outdatedEvents = new ArrayList<>();
-            synchronized (runningJobs) {
-                Iterator<GerritTriggeredEvent> it = runningJobs.iterator();
-                while (it.hasNext()) {
-                    GerritTriggeredEvent runningEvent = it.next();
-
-                    if (runningEvent instanceof ChangeBasedEvent) {
-                        ChangeBasedEvent runningChangeBasedEvent = ((ChangeBasedEvent)runningEvent);
-                        if (!shouldIgnoreEvent(event, policy, runningChangeBasedEvent))
-                        {
-                            outdatedEvents.add(runningChangeBasedEvent);
-                            it.remove();
-                        }
-                    }
-                }
-                // add our new job
-                if (!outdatedEvents.contains(event)) {
-                    runningJobs.add(event);
-                }
-            }
-            // This step can't be done under the lock, because cancelling the jobs needs a lock on higher level.
-            for (ChangeBasedEvent outdatedEvent : outdatedEvents) {
-                logger.debug("Cancelling build for " + outdatedEvent);
-                try {
-                    cancelMatchingJobs(outdatedEvent, jobName);
-                } catch (Exception e) {
-                    // Ignore any problems with canceling the job.
-                    logger.error("Error canceling job", e);
-                }
-            }
-        }
-
-        /**
-         * Determines if event should be ignored due to policy
-         *
-         * @param event event being evaluated
-         * @param policy policy to determine cancellation
-         * @param runningChangeBasedEvent existing event to compare against
-         * @return true if event should be ignored for cancellation
-         */
-        private boolean shouldIgnoreEvent(ChangeBasedEvent event,
-                BuildCancellationPolicy policy, ChangeBasedEvent runningChangeBasedEvent)
-        {
-            // Find all entries in runningJobs with the same Change #.
-            // Optionally, ignore all manual patchsets and don't cancel builds due to
-            // a retrigger of an older build.
-            boolean abortBecauseOfTopic = abortBecauseOfTopic(event,
-                    policy,
-                    runningChangeBasedEvent);
-
-            if (!abortBecauseOfTopic && !runningChangeBasedEvent.getChange().equals(event.getChange())) {
-                return true;
-            }
-
-            boolean shouldCancelManual = (runningChangeBasedEvent instanceof ManualPatchsetCreated
-                    && policy.isAbortManualPatchsets()
-                    || !(runningChangeBasedEvent instanceof ManualPatchsetCreated));
-
-            if (!abortBecauseOfTopic && !shouldCancelManual) {
-                return true;
-            }
-
-            boolean shouldCancelPatchsetNumber = policy.isAbortNewPatchsets()
-                    || Integer.parseInt(runningChangeBasedEvent.getPatchSet().getNumber())
-                    < Integer.parseInt(event.getPatchSet().getNumber());
-
-            if (!abortBecauseOfTopic && !shouldCancelPatchsetNumber) {
-                return true;
-            }
-
-            return false;
-        }
-
-        /**
-         * Tries to cancel any job, which was triggered by the given change event.
-         * <p>
-         * Since the event is always noted in the build cause, it is easy to
-         * identify which specific builds shall be cancelled, without having
-         * to dig down into the parameters, which might've been mutated by the
-         * build while it was running. (This was the previous implementation)
-         * <p>
-         * We look in both the build queue and currently executing jobs.
-         * This extra work is required due to race conditions when calling
-         * Future.cancel() - see
-         * https://issues.jenkins-ci.org/browse/JENKINS-13829
-         *
-         * @param event The event that originally triggered the build.
-         * @param matchOnJobName  job name to match on.
-         */
-        private void cancelMatchingJobs(GerritTriggeredEvent event, String matchOnJobName) {
-            try {
-                if (!(job instanceof Queue.Task)) {
-                    logger.error("Error canceling job. The job is not of type Task. Job name: " + job.getName());
-                    return;
-                }
-
-                // Remove any jobs in the build queue.
-                List<hudson.model.Queue.Item> itemsInQueue = Queue.getInstance().getItems((Queue.Task)job);
-                for (hudson.model.Queue.Item item : itemsInQueue) {
-                    if (checkCausedByGerrit(event, item.getCauses())) {
-                        if (matchOnJobName == null || matchOnJobName.equals(item.task.getName())) {
-                            Queue.getInstance().cancel(item);
-                        }
-                    }
-                }
-
-                String workaround = System.getProperty(JOB_ABORT);
-                if ((workaround != null) && workaround.equals("false")) {
-                    return;
-                }
-
-                // Interrupt any currently running jobs.
-                Jenkins jenkins = Jenkins.getInstance();
-                assert jenkins != null;
-                for (Computer c : jenkins.getComputers()) {
-                    List<Executor> executors = new ArrayList<Executor>();
-                    executors.addAll(c.getOneOffExecutors());
-                    executors.addAll(c.getExecutors());
-                    for (Executor e : executors) {
-                        Queue.Executable currentExecutable = e.getCurrentExecutable();
-
-                        if (currentExecutable != null && currentExecutable instanceof Run<?, ?>) {
-                            Run<?, ?> run = (Run<?, ?>)currentExecutable;
-                            if (checkCausedByGerrit(event, run.getCauses())) {
-                                if (matchOnJobName == null || matchOnJobName.equals(run.getParent().getFullName())) {
-                                    e.interrupt(
-                                            Result.ABORTED,
-                                            new NewPatchSetInterruption()
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-            } catch (Exception e) {
-                // Ignore any problems with canceling the job.
-                logger.error("Error canceling job", e);
-            }
-        }
-
-        /**
-         * Checks if any of the given causes references the given event.
-         *
-         * @param event The event to check for. Checks for <i>identity</i>, not
-         * <i>equality</i>!
-         * @param causes the list of causes. Only {@link GerritCause}s are considered.
-         * @return true if the list of causes contains a {@link GerritCause}.
-         */
-        private boolean checkCausedByGerrit(GerritTriggeredEvent event, Collection<Cause> causes) {
-            for (Cause c : causes) {
-                if (!(c instanceof GerritCause)) {
-                    continue;
-                }
-                GerritCause gc = (GerritCause)c;
-                if (gc.getEvent() == event) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        /**
-         * Removes any reference to the current build for this change.
-         *
-         * @param event the event which started the build we want to remove.
-         * @return true if event was still running.
-         */
-        public boolean remove(ChangeBasedEvent event) {
-            logger.debug("Removing future job " + event.getPatchSet().getNumber());
-            return runningJobs.remove(event);
-        }
-    }
+  
 
     /**
      * Checks that execution must be aborted because of topic.
@@ -2569,7 +2340,7 @@ public class GerritTrigger extends Trigger<Job> {
      * @param runningChange the ongoing change.
      * @return true if so.
      */
-    private boolean abortBecauseOfTopic(ChangeBasedEvent event,
+    protected boolean abortBecauseOfTopic(ChangeBasedEvent event,
                                         BuildCancellationPolicy policy,
                                         ChangeBasedEvent runningChange) {
         String topicName = event.getChange().getTopic();
